Only fetch packages when needed.

+++ scripts/get_debian_source_packages.sh
@@ -8,37 +8,44 @@ SCRIPT_DIR=$(dirname "$0")
 
 set -e
 
-wget --no-check-certificate http://us.archive.ubuntu.com/ubuntu/dists/devel/main/source/Sources.gz
-gzip -d Sources.gz
-mv Sources ubuntu_devel_main_sources.txt
+if [[ ! -f "sources.txt" ]]; then
+    wget --no-check-certificate http://us.archive.ubuntu.com/ubuntu/dists/devel/main/source/Sources.gz
+    gzip -d Sources.gz
+    mv Sources ubuntu_devel_main_sources.txt
 
-wget --no-check-certificate http://us.archive.ubuntu.com/ubuntu/dists/devel/universe/source/Sources.gz
-gzip -d Sources.gz
-mv Sources ubuntu_devel_universe_sources.txt
+    wget --no-check-certificate http://us.archive.ubuntu.com/ubuntu/dists/devel/universe/source/Sources.gz
+    gzip -d Sources.gz
+    mv Sources ubuntu_devel_universe_sources.txt
 
-wget --no-check-certificate http://us.archive.ubuntu.com/ubuntu/dists/devel/multiverse/source/Sources.gz
-gzip -d Sources.gz
-mv Sources ubuntu_devel_multiverse_sources.txt
+    wget --no-check-certificate http://us.archive.ubuntu.com/ubuntu/dists/devel/multiverse/source/Sources.gz
+    gzip -d Sources.gz
+    mv Sources ubuntu_devel_multiverse_sources.txt
 
-wget --no-check-certificate https://ftp.nl.debian.org/debian/dists/sid/main/source/Sources.gz
-gzip -d Sources.gz
-mv Sources debian_sid_main_sources.txt
+    wget --no-check-certificate https://ftp.nl.debian.org/debian/dists/sid/main/source/Sources.gz
+    gzip -d Sources.gz
+    mv Sources debian_sid_main_sources.txt
 
-wget --no-check-certificate https://repo.pureos.net/pureos/dists/green/main/source/Sources.xz
-unxz -d Sources.xz
-mv Sources pureos_green_main_sources.txt
+    wget --no-check-certificate https://repo.pureos.net/pureos/dists/green/main/source/Sources.xz
+    unxz -d Sources.xz
+    mv Sources pureos_green_main_sources.txt
 
-wget --no-check-certificate https://repo.pureos.net/pureos/dists/landing/main/source/Sources.xz
-unxz -d Sources.xz
-mv Sources pureos_landing_main_sources.txt
+    wget --no-check-certificate https://repo.pureos.net/pureos/dists/landing/main/source/Sources.xz
+    unxz -d Sources.xz
+    mv Sources pureos_landing_main_sources.txt
 
-wget --no-check-certificate https://repo.pureos.net/pureos/dists/amber/main/source/Sources.xz
-unxz -d Sources.xz
-mv Sources pureos_amber_main_sources.txt
+    wget --no-check-certificate https://repo.pureos.net/pureos/dists/amber/main/source/Sources.xz
+    unxz -d Sources.xz
+    mv Sources pureos_amber_main_sources.txt
 
-wget --no-check-certificate https://repo.pureos.net/pureos/dists/byzantium/main/source/Sources.xz
-unxz -d Sources.xz
-mv Sources pureos_byzantium_main_sources.txt
+    wget --no-check-certificate https://repo.pureos.net/pureos/dists/byzantium/main/source/Sources.xz
+    unxz -d Sources.xz
+    mv Sources pureos_byzantium_main_sources.txt
 
-cat *_sources.txt > sources.txt
-rm *_sources.txt
+    cat *_sources.txt > sources.txt
+    rm *_sources.txt
+fi
+
+for line in sources.txt
+do
+    echo "$line"
+done
